Extract pure UI helpers into web/js/ui/ modules
Move format/encoding/environment helpers into web/js/ui/format.js and
channel cell-value coercion into web/js/ui/channel-values.js. Both are
free of UI state and the document, so they are directly unit-testable.

ui.js re-exports buildExportFileName to keep its public entry point
stable for existing importers and tests. No user-visible behavior change.
